refactor(megane-core): consolidate element data and split large parse functions
- Add atomic.rs centralizing all element data (symbol↔atomic number,
  covalent/VdW radii, masses, atom-name heuristics) previously
  duplicated across parser.rs, bonds.rs, lammps_data.rs, and gro.rs
- Add trajectory.rs with shared TrajectoryData struct; XtcData and
  LammpstrjData become type aliases preserving backwards compatibility
- Split lammps_data::parse() (306-line monolith) into focused helpers:
  parse_header, parse_masses_section, parse_atoms_section,
  parse_bonds_section — each independently testable
- Migrate mol.rs error tests from panic!("expected error") to
  assert!(result.is_err()) with descriptive messages
- Deduplicate ~250 lines of element/radius/mass tables
- All 70 Rust tests pass; WASM build and 266 TypeScript tests pass
