+++ crates/megane-core/src/lib.rs
@@ -1,3 +1,4 @@
+pub mod atomic;
 pub mod bonds;
 pub mod cif;
 pub mod gro;
@@ -7,7 +8,9 @@ pub mod mol;
 pub mod parser;
 pub mod top;
 pub mod traj;
+pub mod trajectory;
 pub mod xtc;
 pub mod xyz;
 
 pub use parser::ParsedStructure;
+pub use trajectory::TrajectoryData;
