+++ crates/megane-core/src/parser.rs
@@ -3,6 +3,10 @@
 /// Handles ATOM/HETATM, CRYST1, CONECT, and MODEL/ENDMDL records.
 use std::collections::{HashMap, HashSet};
 
+// Re-export element utilities that were historically part of this module.
+// New code should import from `crate::atomic` directly.
+pub use crate::atomic::{capitalize, symbol_to_atomic_num};
+
 /// Parsed atom data from a single ATOM/HETATM line.
 pub struct Atom {
     pub x: f32,
@@ -22,93 +26,6 @@ pub struct ParsedStructure {
     pub box_matrix: Option<[f32; 9]>,
     pub frame_positions: Vec<Vec<f32>>,
     pub atom_labels: Option<Vec<String>>,
-}
-
-/// Element symbol → atomic number lookup.
-pub fn symbol_to_atomic_num(sym: &str) -> u8 {
-    match sym {
-        "H" => 1,
-        "He" => 2,
-        "Li" => 3,
-        "Be" => 4,
-        "B" => 5,
-        "C" => 6,
-        "N" => 7,
-        "O" => 8,
-        "F" => 9,
-        "Ne" => 10,
-        "Na" => 11,
-        "Mg" => 12,
-        "Al" => 13,
-        "Si" => 14,
-        "P" => 15,
-        "S" => 16,
-        "Cl" => 17,
-        "Ar" => 18,
-        "K" => 19,
-        "Ca" => 20,
-        "Sc" => 21,
-        "Ti" => 22,
-        "V" => 23,
-        "Cr" => 24,
-        "Mn" => 25,
-        "Fe" => 26,
-        "Co" => 27,
-        "Ni" => 28,
-        "Cu" => 29,
-        "Zn" => 30,
-        "Ga" => 31,
-        "Ge" => 32,
-        "As" => 33,
-        "Se" => 34,
-        "Br" => 35,
-        "Kr" => 36,
-        "Rb" => 37,
-        "Sr" => 38,
-        "Y" => 39,
-        "Zr" => 40,
-        "Mo" => 42,
-        "Ru" => 44,
-        "Rh" => 45,
-        "Pd" => 46,
-        "Ag" => 47,
-        "Cd" => 48,
-        "In" => 49,
-        "Sn" => 50,
-        "Sb" => 51,
-        "Te" => 52,
-        "I" => 53,
-        "Xe" => 54,
-        "Cs" => 55,
-        "Ba" => 56,
-        "La" => 57,
-        "Ce" => 58,
-        "Nd" => 60,
-        "Sm" => 62,
-        "Eu" => 63,
-        "Gd" => 64,
-        "Tb" => 65,
-        "Dy" => 66,
-        "Ho" => 67,
-        "Er" => 68,
-        "Yb" => 70,
-        "Lu" => 71,
-        "Hf" => 72,
-        "Ta" => 73,
-        "W" => 74,
-        "Re" => 75,
-        "Os" => 76,
-        "Ir" => 77,
-        "Pt" => 78,
-        "Au" => 79,
-        "Hg" => 80,
-        "Tl" => 81,
-        "Pb" => 82,
-        "Bi" => 83,
-        "Th" => 90,
-        "U" => 92,
-        _ => 0,
-    }
 }
 
 /// Parse the element from a PDB ATOM/HETATM line.
@@ -154,18 +71,6 @@ fn parse_element(line: &str) -> u8 {
     }
 
     0 // unknown
-}
-
-pub fn capitalize(s: &str) -> String {
-    let mut chars = s.chars();
-    match chars.next() {
-        None => String::new(),
-        Some(c) => {
-            let upper: String = c.to_uppercase().collect();
-            let lower: String = chars.flat_map(|c| c.to_lowercase()).collect();
-            format!("{}{}", upper, lower)
-        }
-    }
 }
 
 /// Parse an ATOM/HETATM line to extract coordinates and element.
@@ -421,31 +326,6 @@ pub fn parse(text: &str) -> Result<ParsedStructure, String> {
 #[cfg(test)]
 mod tests {
     use super::*;
-
-    #[test]
-    fn test_symbol_to_atomic_num_common() {
-        assert_eq!(symbol_to_atomic_num("H"), 1);
-        assert_eq!(symbol_to_atomic_num("C"), 6);
-        assert_eq!(symbol_to_atomic_num("N"), 7);
-        assert_eq!(symbol_to_atomic_num("O"), 8);
-        assert_eq!(symbol_to_atomic_num("S"), 16);
-        assert_eq!(symbol_to_atomic_num("Fe"), 26);
-        assert_eq!(symbol_to_atomic_num("Na"), 11);
-    }
-
-    #[test]
-    fn test_symbol_to_atomic_num_unknown() {
-        assert_eq!(symbol_to_atomic_num("Xx"), 0);
-        assert_eq!(symbol_to_atomic_num(""), 0);
-    }
-
-    #[test]
-    fn test_capitalize() {
-        assert_eq!(capitalize("fe"), "Fe");
-        assert_eq!(capitalize("FE"), "Fe");
-        assert_eq!(capitalize("c"), "C");
-        assert_eq!(capitalize(""), "");
-    }
 
     #[test]
     fn test_parse_minimal_pdb() {
